Create serverController.js
Moved server function code to serverController.js
